modified:   MNISTTool/LinearRegressionMNIST.java 	modified:   MNISTTool/MakeTheta.java 	modified:   MNISTTool/TrainData.java 	modified:   ../webapp/WEB-INF/views/home.jsp 	modified:   ../../../target/m2e-wtp/web-resources/META-INF/maven/org.taek/controller/pom.properties

+++ src/main/java/MNISTTool/LinearRegressionMNIST.java
@@ -23,7 +23,7 @@ public class LinearRegressionMNIST implements LinearRegression {
 		System.out.println("h : " + h);
 	}
 	@Override
-	public Long costFuntion() {
+	public Double costFuntion() {
 		return null;
 	}
 	@Override
